Add helpers for an improved Session base class.
Each new Session subclass must re-implement cumbersome techniques
to track the state of an SMTK model relative to the modeling kernel.
This commit provides an implementation of transcribeInternal that
base classes may use as an alternative to some of this recordkeeping.
If a Session does not override transcribeInternal, it must implement
some simple new methods queue related entities when requested by the
base class.