Fix `Manager::findCellHasUseWithSense()` method.
It needs the use's UUID to resolve ambiguities with multiple orientations.
Another method should be created (`findCellUsesWithSense`?) that
can return multiple arrangements when no use is given
(either because it is unknown or because the use is a vertex-use
record that may have multiple use records with the same sense
and orientation).