Progress on attribute system callbacks.
An initial implementation of callbacks for attribute-system events.
Each event has its own special class for reporting event state.

Although not implemented yet, the plan is to provide an "observe"
method on the attribute system so that calling
  `sys->observe([](const XXXEvent&) { ... });` on an attribute
system will invoke your lambda when the given event occurs.

Internally, this is accomplished by calling
  `this->trigger(XXXEvent(...));`
from within methods on the System (or related classes) just before or
after the event.

Some more work is required to
+ implement more event classes.
+ deal with high-level events (such as a callback invoked whenever
  *any* event on an attribute *or* its children occurs).
+ benchmark the overhead of the events on callbacks.